ANDROID: Export the symbol of ext_sched_class
The function task_on_scx() that determines whether a task is an SCX task
depends on the symbol ext_sched_class. Some scheduling-related modules
also need to use task_on_scx() to determine which scheduling class a
task belongs to, so the symbol ext_sched_class need to be exported.

Bug: 417106548
Bug: 430229917
Change-Id: I50ab7e47850f719c0d83dd4d0096843f294ff580
Signed-off-by: Tengfei Fan <quic_tengfan@quicinc.com>
Signed-off-by: Zhou Jian <quic_jianzhou@quicinc.com>