ANDROID: GKI: Fix up abi issue in struct scm_fp_list
In commit e7026f590200 ("UPSTREAM: af_unix: Try to run GC async."), a
new field, count_unix, was added to the structure.  This structure is
ONLY dynamically allocated by the core kernel, and only accessable by a
pointer from anyone else.  To work around abi issues, move the new field
to the end of the structure, ensuring that anyone who was attempting to
access the fields of the structure, would continue to do so in a safe
way.

Add some __GENKSYMS__ magic to keep the crc generation sane, but the
field still needs to be documented as an abi break, so add it to the
abi_gki_aarch64.stg.allowed_breaks file.

Bug: 404256079
Change-Id: I849000eac4f9001b04e4d110dfa74c63067cb786
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Fixes: e7026f590200 ("UPSTREAM: af_unix: Try to run GC async.")
(cherry picked from commit 452d899d2f6ba7fa4085b17495bd8402005715cf)
Signed-off-by: Lee Jones <joneslee@google.com>