Revert "ANDROID: mm: Set PAGE_BLOCK_ORDER to 8 when ARM64_16K_PAGES"
This reverts commit 45afa562802ae9a253ef437dd0010c8c2ec17806.

Reason for revert: This was a workaround due to the kernel build
tools preserving the # nocheck comment as part of the config option
for CONFIG_PAGE_BLOCK_ORDER, which is problematic, since it is
supposed to be an int. The build tools have been patched to not
do that anymore, so this can be removed.

Bug: 424212284
Bug: 375647879
Bug: 355449177
Bug: 418282543
Change-Id: Ib999aea150c1c5f7f22ea6bdd81de0ec75f8efaf
[isaacmanjarres: resolve merge conflicts due to PAGE_BLOCK_ORDER
being hardcoded to 8 on android14 kernels.]
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>