ANDROID: add task_dma_buf_record and allocation preload facilities
The records for task's dma buffers will need to be pre-allocated in
advance, therefore a mechanism is needed to allocate and store them
before they are used. Introduce task_dma_buf_record struct that will
be used in follow-up patches to store task to dmabuf association and
implement per-cpu storage for task_dma_buf_records.

Bug: 424648392
Bug: 441733128
Suggested-by: Minchan Kim <minchan@google.com>
Change-Id: I9009a668c22ba4464347385451c6fb65f7d83af7
Signed-off-by: Suren Baghdasaryan <surenb@google.com>