ANDROID: Build fixups with PROXY_EXEC v18 + !CONFIG_SMP
In doing some testing, I caught a few build issues with
different combinations of enabling and disabling
CONFIG_SCHED_PROXY_EXEC and CONFIG_SMP after syncing
with the v18 of the proxy-exec series.

These fixes were not in the v18 series of the patch, but
I've included them in my tree for v19.

Bug: 427820735
Change-Id: Ice3733219a292f882223f8a505f5339fa4b7583f
Signed-off-by: John Stultz <jstultz@google.com>