ANDROID: Fixup init_user_ns CRC change
The previous patch, by including the new trace/hooks/users.h file
causes other heads to be included which sligtly change the
declaration typing info for a number types to not be opaque types.

This results in the struct user_namespace init_user_ns value
to get a different CRC result.

To address this, use the ANDROID_KABI_DECLONLY macro to clarify
the affected types are only considered to be delared.

This can hopefully be dropped during a future KABI thaw.

Bug: 425550656
Change-Id: I18c463535b04e766777f7ebb8115fedf8e8f0c00
Signed-off-by: John Stultz <jstultz@google.com>