ANDROID: rust_binder: release threads before refs
Commit 5da94eacba68 ("ANDROID: rust_binder: print and avoid underflow
with invalid refcnt") added a print when decrementing the refcount if
the ref doesn't exist. This revealed that this sometimes happens inside
thread.release() because the refs were cleared before releasing the
threads. In principle this error is harmless, but to avoid printing this
error, release threads before clearing refs.

Bug: 430660253
Change-Id: I9ceb71ca831f5bb58193e12951245bab19ea4fa2
Signed-off-by: Alice Ryhl <aliceryhl@google.com>