ANDROID: mm: Added memcg1_charge_batch
Added a new function memcg1_charge_batch, which is called by the
vendor's ko. It is similar to memcg1_commit_charge, but charges a batch
of folios at a time.
The reason for introducing it is that for a batch of folios, each call
to memcg1_commit_charge will turn off and on interrupts, which will
cause a large performance degradation.

Bug: 431672372
Bug: 435545311
Change-Id: I91712d02b7f119b7f35d33503c172d9753a9a69e
Signed-off-by: Pengfei Li <pengfei.kernel@vivo.corp-partner.google.com>
(cherry picked from commit c04583979e3215e4df6025701502a3e638d25999)