ANDROID: BACKPORT: vsock/virtio: Limit size of TX buffers
Prior to I4212a8daf9f19b5bbffc06ce93338c823de7bb19 ("FROMGIT:
vhost/vsock: Allocate nonlinear SKBs for handling large receive
buffers"), vhost would allocate a linear SKB of up to 128KiB in size,
limited by VIRTIO_VSOCK_MAX_PKT_BUF_SIZE and the length of the incoming
packet.

As such large allocations are prone to failure, an Android-specific
command-line option to reduce VIRTIO_VSOCK_MAX_PKT_BUF_SIZE was added in
Ia4f9180025b64d61d9e2b882daa84460b8f08fd2 ("ANDROID: Make vsock virtio
packet buff size configurable"). Since then, Cuttlefish has been
specifying this as 16Kib for its guest kernel images, resulting in a
maximum RX buffer allocation of 32KiB (which correlates with
PAGE_ALLOC_COSTLY_ORDER for architectures using 4KiB pages) on the host.

Now that both the vhost and virtio backends for vsock support non-linear
SKB allocations on the transmit and receive paths, we are guaranteed not
to allocate above PAGE_ALLOC_COSTLY_ORDER and so the Android-specific
command-line tunable can be removed.

Well, almost.

The fly in the ointment is that we are unable to update both the guest
and host kernels atomically and, in the case of Cuttlefish, the host
kernel may be of a sufficient vintage that backporting the support
for non-linear SKBs in vsock is non-trivial and impractical.

In an effort to phase out the command-line option but preserve existing
behaviour with older hosts, limit the TX buffer size to ~28KiB. This
means that the unpatched host will continue to allocate 32KiB RX buffers
but will achieve better utilisation, whilst having a page of wiggle room
in case the size of an SKB increases before it learns about non-linear
allocations. This additionally saves vsock users in AVF from having to
worry about whether or not they need to configure the buffer size
manually.

One day, when all supported hosts are running kernels later than
android16-6.12, we'll be able to revert this. In the meantime, it's a
bit of a mess.

Bug: 421244320
Bug: 149494272
[willdeacon@: Drop zero-copy part]
Signed-off-by: Will Deacon <willdeacon@google.com>
Change-Id: Iae8075ce39b3de6471f875e06869e4899f6eed5a